maggotuba/requirements.txt
absl-py==0.15.0
astunparse==1.6.3
attrdict==2.0.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
flatbuffers==1.12
fonttools==4.31.2
gast==0.4.0
geomloss==0.2.5
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
gudhi==3.5.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.11.3
joblib==1.1.0
keopscore==2.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
llvmlite==0.38.0
Markdown==3.3.6
mat73==0.58
matplotlib==3.5.1
numba==0.55.1
numpy==1.19.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
plotly==5.7.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.2
pykeops==2.0
pynndescent==0.5.6
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
six==1.15.0
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
threadpoolctl==3.1.0
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.0
typing-extensions==3.7.4.3
umap-learn==0.5.2
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.12.1
zipp==3.8.0
